Fixed a table mistakenly created FK from the messages table. There was a column named "sender_id" which referenced the messages table itself.
In addition, I changed the same column name to "admins_id" in the DB and
in the Message model.
I changed its FK reference to the admin's table because the users are already
referenced in the feedback table (related to the messages table).

Deleted insertion of sender_id in the 'FeedbacksController' because it no longer exists
AND there is no need for this considering message already knows which user sent it
because it is linked to the feedbacks table.